services/data-acquisition/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket for MCP
websockets==12.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
pytz==2023.3

# Configuration and Environment
python-dotenv==1.0.0

# Type Hints
typing-extensions==4.12.0

# Utilities
python-dateutil==2.8.2

# Optional: Database (for later phases)
# asyncpg==0.29.0
# psycopg2-binary==2.9.9

# Optional: Redis (for later phases)  
# redis==5.0.1